Rearranging tests (5)
